feat: hydra-cli refactoring and fixes after review (rename commands, graceful shutdowns)
affects: @subsquid/hydra-cli

hydra-cli refactoring and fixes after review (rename commands, graceful shutdowns)
