+++ packages/hydra-cli/package.json
@@ -52,6 +52,7 @@
     "@octokit/auth-oauth-device": "^3.1.2",
     "@subsquid/openreader": "^0.2.0",
     "chalk": "^4.1.2",
+    "cli-select": "^1.1.2",
     "cli-ux": "^5.6.3",
     "dotenv": "^10.0.0",
     "execa": "^4.1.0",
