+++ packages/hydra-cli/src/commands/app/deploy.ts
@@ -1,8 +1,12 @@
 import { Command, flags } from '@oclif/command'
-import simpleGit, { SimpleGit, SimpleGitOptions } from 'simple-git'
+import simpleGit, {
+  RemoteWithRefs,
+  SimpleGit,
+  SimpleGitOptions,
+} from 'simple-git'
 import Debug from 'debug'
-import { deploy } from '../rest-client/routes/deploy'
-import { cli } from 'cli-ux'
+import { deploy } from '../../rest-client/routes/deploy'
+import cliSelect from 'cli-select'
 
 const debug = Debug('qnode-cli:deploy')
 const options: Partial<SimpleGitOptions> = {
@@ -27,19 +31,29 @@ export default class Deploy extends Command {
     debug(`Parsed flags: ${JSON.stringify(flags, null, 2)}`)
     const deploymentName = flags.name
 
+    let remoteUrl: RemoteWithRefs
     const remotes = await git.getRemotes(true)
-    let remotesCLIList = ''
-    for (let i = 1; i <= remotes.length; i++) {
-      remotesCLIList += `\n${i}. ${remotes[i - 1].name}`
+    if (remotes.length === 0) {
+      this.error(`The remotes were not found`, { code: '1' })
+    } else if (remotes.length === 1) {
+      remoteUrl = remotes[0]
+    } else {
+      const selected = await cliSelect({
+        cleanup: false,
+        values: remotes.map((remote) => remote.name),
+      }).catch(() => {
+        this.error('Canceled', { code: '1' })
+      })
+      remoteUrl = remotes.find(
+        (remote) => remote.name === selected.value
+      ) as RemoteWithRefs
     }
-    const remoteNumber = await cli.prompt(
-      `Enter corresponding remote number: ${remotesCLIList}\n`,
-      { required: true }
-    )
-    const remoteUrl = remotes[Number(remoteNumber) - 1]
-    if (remoteUrl === undefined) {
-      throw Error(`Incorrect input`)
-    }
+    await git.listRemote([remoteUrl.name]).catch(() => {
+      this.error(`Remote url with name ${remoteUrl.name} not exists`, {
+        code: '1',
+      })
+    })
+
     const currentBranch = (await git.branch()).current
 
     // show warning if head origin commit is not the same as the
@@ -48,7 +62,7 @@ export default class Deploy extends Command {
     const diffs = await git.diffSummary([
       currentBranch,
       `${remoteUrl.name}`,
-      `${currentBranch}`, // TODO check on real case!
+      `${currentBranch}`,
     ])
     if (
       (status.files && status.files.length) ||
@@ -60,16 +74,19 @@ export default class Deploy extends Command {
       this.warn(
         `Head origin commit is not the same as the local origin commit or there are unstaged or uncomitted changes`
       )
-      const enterResult = await cli.prompt(
-        `Are you sure (enter corresponding remote number): \n1.yes\n[any other key].no\n`,
-        { required: true }
-      )
-      if (enterResult !== '1') {
-        this.log('Canceled')
-        return
+      this.log(`Are you sure?`)
+      const selected = await cliSelect({
+        cleanup: false,
+        values: ['yes', 'no'],
+      }).catch(() => {
+        this.error('Canceled', { code: '1' })
+      })
+      if (selected.value === 'no') {
+        this.error('Canceled', { code: '1' })
       }
     }
 
+    this.log(`🦑 Releasing the Squid at ${remoteUrl.name}`)
     const message = await deploy(
       deploymentName,
       `${remoteUrl.refs.fetch}#${currentBranch}`
